Address EBUSY failures when running using NFS mounts for the scratch folder.

+++ NextGen_Forcings_Engine_BMI/NextGen_Forcings_Engine/bmi_model.py
@@ -1,6 +1,7 @@
 # Need these for BMI
 # This is needed for get_var_bytes
 import os
+import gc
 from pathlib import Path
 
 import netCDF4 as nc
@@ -719,31 +720,31 @@ class NWMv3_Forcing_Engine_BMI_model(Bmi):
         :return: None
         """
 
+        # Force destruction of ESMF objects
+        self._WrfHydroGeoMeta = None
+        self._inputForcingMod = None
+        self._suppPcpMod = None
+        self._model = None
+
+        # Try moving this after all of the ESMF and model bits have
+        # been disposed of - maybe they were keeping something open.
+        #
+        # Potential workaround if that's not enough: uncomment the
+        # return before the file cleanup block, leak the files during
+        # the job, and let the workflow clean them up after the
+        # process exits
+        gc.collect()  # make sure objects are deleted from memory
         if self._mpi_meta.rank == 0:
             for filename in os.listdir(self._job_meta.scratch_dir):
-                file_path = os.path.join(self._job_meta.scratch_dir, filename)
-                if os.path.isfile(file_path) and filename[0:23] != "NextGen_Forcings_Engine":
-                    os.remove(file_path)
-                elif os.path.isdir(file_path):
-                    os.rmdir(file_path)
-
-        # Force destruction of ESMF objects
-        try:
-            del self._WrfHydroGeoMeta
-        except AttributeError:
-            pass
-
-        try:
-            del self._inputForcingMod
-        except AttributeError:
-            pass
-
-        try:
-            del self._suppPcpMod
-        except AttributeError:
-            pass
-
-        self._model = None
+                # NFS mounts may create temporary files to facilitate read-after-delete functionality on linux systems
+                # these will be cleaned when the mount is removed but will throw an error if python tries to remove it
+                # the file name is typically ".nfs" followed by numbers, so we'll just ignore files that start with it
+                if not filename.startswith(".nfs"):
+                    file_path = os.path.join(self._job_meta.scratch_dir, filename)
+                    if os.path.isfile(file_path) and filename[0:23] != "NextGen_Forcings_Engine":
+                        os.remove(file_path)
+                    elif os.path.isdir(file_path):
+                        os.rmdir(file_path)
 
     # -------------------------------------------------------------------
     # -------------------------------------------------------------------
